Docs: Drop test workbench list
I'll try to test as many desktop environments and distro combos as
possible prior to any release that may bring breaking changes. As a solo
developer, it is a little too much to make a commitment to test all 10
environment combos. The packager would normally report any breakages.

A friend has offered to assist with QA testing before the v1.0.0
release, but in the long run, since this list is just another thing to
maintain, it will be dropped.