Workaround Jekyll quirk with 'inherits' pages
Due to possibly the way Jekyll renders pages, or it might be a bug.

Rename ubuntu.md -> -ubuntu.md so it is rendered first. The logic where
derivative pages 'inherits' another page seem to be rendering with
literal {% %}'s. Except for Zorin OS, so it can only be explained by the
filename order.